[CH] Reduce emoji mart css footprint
