--- app/soapbox/features/emoji/emoji_picker.js
@@ -1,7 +1,0 @@
-import Emoji from 'emoji-mart/dist-es/components/emoji/emoji';
-import Picker from 'emoji-mart/dist-es/components/picker/picker';
-
-export {
-  Picker,
-  Emoji,
-};
